clean out all the old stuff
